IGNITE-2805: Changed transactions retrieving logic in IgniteImpl and added more test.

+++ modules/platforms/cpp/jni/src/java.cpp
@@ -992,7 +992,7 @@ namespace ignite
                         {
                             GetJniErrorMessage(errMsg, res);
 
-                            errMsgLen = errMsg.length();
+                            errMsgLen = static_cast<int>(errMsg.length());
                         }
                     }
